% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Dummy data for TomoQC example}
\format{
An object of class \code{list} of length 3.
}
\usage{
example_data
}
\description{
Dummy data for TomoQC example
}
\author{
Erik Schild \email{erikschild@live.nl}
}
\keyword{datasets}
