% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.function.only.R
\name{tomo_diagnostic}
\alias{tomo_diagnostic}
\title{tomo_diagnostic}
\usage{
tomo_diagnostic(
  transcripts,
  reads,
  umis,
  plot_title = "QC plots",
  cutoff_spike = 0.25,
  cutoff_genes = 2000,
  spike_ins = T,
  barcodes = c(1:96)
)
}
\arguments{
\item{transcripts}{data.frame containing transcript counts (reads corrected using UMIs), with gene names as rownames}

\item{reads}{data.frame containing read counts, with gene names as rownames}

\item{umis}{data.frame containing UMI counts, with genen names as rownames}

\item{plot_title}{a string to title the QC plot}

\item{cutoff_spike}{a cutoff value for spike-in ratio (ranges from 0 to 1)}

\item{cutoff_genes}{a cutoff value for minimum unique gene count per slice}

\item{spike_ins}{Controls whether spike-ins were used, and thus if a spike-ins percentage plot should be generated. Default = TRUE}

\item{barcodes}{this is probably gone in the final version}
}
\value{
Provides various QC plots to quickly assess the quality of tomo-seq data
}
\description{
tomo_diagnostic
}
