% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.function.only.R
\name{tomo_quality}
\alias{tomo_quality}
\title{Tomosequencing quality control function}
\usage{
tomo_quality(
  transcripts = example_data$transcripts,
  reads = example_data$reads,
  umis = example_data$UMIs,
  plot_title = "QC plots",
  cutoff_spike = 25,
  cutoff_genes = 100,
  spike_ins = T
)
}
\arguments{
\item{transcripts}{data.frame containing transcript counts, with gene names in the first column}

\item{reads}{data.frame containing read counts, with gene names in the first column}

\item{umis}{data.frame containing UMI counts, with genen names in the first column}

\item{plot_title}{a string to title the QC plot}

\item{cutoff_spike}{a cutoff value for spike-in percentage of total Transcripts (ranges from 0 to 100)}

\item{cutoff_genes}{a cutoff value for minimum unique gene count per slice}

\item{spike_ins}{Controls whether spike-ins were used, and thus if a spike-ins percentage plot should be generated. Default = TRUE}
}
\value{
Provides various QC plots to quickly assess the quality of tomo-seq data
}
\description{
Tomosequencing quality control function
}
